Refactor to follow clean architecture
