--- core/src/main/java/Entity/BlockType.java
@@ -1,9 +1,0 @@
-package Entity;
-
-public enum BlockType {
-    AIR,
-    GRASS,
-    DIRT,
-    STONE,
-    WATER
-}
