--- core/src/main/java/Entity/Item.java
@@ -1,58 +1,0 @@
-package Entity;
-
-import java.util.Objects;
-
-public class Item {
-
-    private final String id;
-    private final String displayName;
-    private final boolean stackable;
-
-    public Item(String id, String displayName, boolean stackable) {
-        if (id == null || id.isEmpty()) {
-            throw new IllegalArgumentException("Item ID must not be empty");
-        }
-        if (displayName == null || displayName.isEmpty()) {
-            throw new IllegalArgumentException("Display name must not be empty");
-        }
-
-        this.id = id;
-        this.displayName = displayName;
-        this.stackable = stackable;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-
-    public boolean isStackable() {
-        return stackable;
-    }
-
-    /**
-     * Determines whether this item is equal to another object.
-     * @param o The object to compare with this item.
-     * @return True if the specified object is an Item with the same ID; false otherwise.
-     */
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Item)) return false;
-        Item item = (Item) o;
-        return id.equals(item.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-
-    @Override
-    public String toString() {
-        return displayName + " (" + id + ")";
-    }
-}
