--- core/src/main/java/Entity/Sphere.java
@@ -1,12 +1,0 @@
-package Entity;
-
-import com.badlogic.gdx.graphics.g3d.Model;
-import com.badlogic.gdx.graphics.g3d.ModelInstance;
-
-public class Sphere extends ModelInstance {
-
-    public Sphere(Model model){
-        super(model);
-    }
-
-}
