--- core/src/main/java/InputBoundary/FirstPersonCameraController.java
@@ -1,10 +1,0 @@
-package InputBoundary;
-
-import Entity.Player;
-import com.badlogic.gdx.graphics.Camera;
-
-public class FirstPersonCameraController extends CameraController {
-    public FirstPersonCameraController(Camera camera, Player player) {
-        super(camera, player);
-    }
-}
