--- core/src/main/java/UseCases/PlayerMovement/PlayerMovementInputBoundary.java
@@ -1,5 +1,0 @@
-package UseCases.PlayerMovement;
-
-public interface PlayerMovementInputBoundary {
-    void execute(PlayerMovementInputData playerMovementInputData, float deltaTime);
-}
