+++ core/src/main/java/application/use_cases/ChunkGeneration/ChunkGenerationInputBoundary.java
@@ -1,0 +1,5 @@
+package application.use_cases.ChunkGeneration;
+
+public interface ChunkGenerationInputBoundary {
+    void execute(ChunkGenerationInputData inputData);
+}
