+++ core/src/main/java/application/use_cases/PlayerMovement/PlayerMovementInputBoundary.java
@@ -1,0 +1,5 @@
+package application.use_cases.PlayerMovement;
+
+public interface PlayerMovementInputBoundary {
+    void execute(PlayerMovementInputData playerMovementInputData);
+}
