+++ core/src/main/java/domain/entities/BlockType.java
@@ -1,0 +1,9 @@
+package domain.entities;
+
+public enum BlockType {
+    AIR,
+    GRASS,
+    DIRT,
+    STONE,
+    WATER
+}
