--- core/src/main/java/io/github/testlibgdx/CubeRender.java
@@ -1,30 +1,0 @@
-package io.github.testlibgdx;
-
-import com.badlogic.gdx.graphics.g3d.Model;
-import com.badlogic.gdx.graphics.g3d.ModelInstance;
-import com.badlogic.gdx.math.Vector3;
-
-
-// TODO: Similar to CubeFactory also currently useless
-public class CubeRender extends ModelInstance {
-    private final Vector3 position = new Vector3();
-    public CubeRender(Model model) {
-        super(model);
-    }
-
-    public float getX() {
-        super.transform.getTranslation(position);
-        return position.x;
-    }
-
-    public float getY() {
-        super.transform.getTranslation(position);
-        return position.y;
-    }
-
-    public float getZ() {
-        super.transform.getTranslation(position);
-        return position.z;
-    }
-
-}
