+++ core/src/main/java/presentation/controllers/FirstPersonCameraController.java
@@ -1,0 +1,10 @@
+package presentation.controllers;
+
+import domain.entities.Player;
+import com.badlogic.gdx.graphics.Camera;
+
+public class FirstPersonCameraController extends CameraController {
+    public FirstPersonCameraController(Camera camera, Player player) {
+        super(camera, player);
+    }
+}
