+++ core/src/main/java/presentation/view/ViewManager.java
@@ -1,0 +1,17 @@
+package presentation.view;
+
+public class ViewManager {
+    private Viewable mainView;
+    public ViewManager() {
+        mainView = new GameView();
+        mainView.createView();
+    }
+
+    public void render() {
+        if (mainView != null) mainView.renderView();
+    }
+
+    public void dispose() {
+        if (mainView != null) mainView.disposeView();
+    }
+}
