+++ core/src/main/java/presentation/view/Viewable.java
@@ -1,0 +1,7 @@
+package presentation.view;
+
+public interface Viewable {
+    void createView();
+    void renderView();
+    void disposeView();
+}
