--- core/src/main/java/view/Viewable.java
@@ -1,7 +1,0 @@
-package view;
-
-public interface Viewable {
-    void createView();
-    void renderView();
-    void disposeView();
-}
